ProductController.java added and a jsp to show results

+++ src/main/java/com/emag/model/dao/IProductDAO.java
@@ -7,8 +7,8 @@ import com.emag.model.Product;
 public interface IProductDAO {
 	
 	void addProduct(Product product) throws Exception;
-	void deleteProduct(int product_id) throws Exception;
+	void deleteProduct(long product_id) throws Exception;
 	void updateProduct(Product product) throws Exception;
-	Product getProductById(int product_id) throws Exception;
+	Product getProductById(long product_id) throws Exception;
 	List<Product> getProductsByCategory(int category_id) throws Exception;
 }
